scripts/quickstart: Fix passing -d DOMAIN to scripts/make-env
Fix getopts so that scripts/quickstart -d takes the DOMAIN arg:

Without it, d) DOMAIN="${OPTARG}" is empty and make-env aborts
because DOMAIN is empty.

Update make-env to write this error to stderr so it can be seen
on the shell (scripts/quickstart redirects its stdout to a file).

Also check the return value make-env and only overwrite the config
in ${CONFIG_DIR}/activate when make-env didn't return an error.

This fixes the documented use of quickstart -d to set the domain.